+++ cache_alchemy/backends/base.py
@@ -3,17 +3,15 @@ from abc import ABC, abstractmethod
 from types import FunctionType
 from typing import (
     Any,
-    Callable,
     ContextManager,
     Dict,
     Tuple,
-    TypeVar,
     cast,
     Pattern,
-    Set,
-    Optional,
 )
+from typing import Callable, TypeVar, Optional, Set, Generic
 
+from ..config import DefaultConfig
 from ..utils import (
     generate_strict_key,
     generate_fast_key,
@@ -25,7 +23,7 @@ ReturnType = TypeVar("ReturnType")
 CacheFunctionType = Callable[..., ReturnType]
 
 
-class BaseCache(ABC):
+class BaseCache(Generic[ReturnType], ABC):
     def __init__(
         self,
         *,
@@ -118,3 +116,97 @@ class BaseCache(ABC):
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         pass
+
+
+DistributedCacheReturnType = TypeVar("DistributedCacheReturnType", bound=bytes)
+
+
+class DistributedCache(BaseCache[DistributedCacheReturnType]):
+    def __init__(self, *, cached_function: FunctionType, **kwargs):
+        super().__init__(cached_function=cached_function, **kwargs)
+        self.client = DefaultConfig.get_current_config().cache_redis_client
+        if self.client.connection_pool.connection_kwargs.get("decode_responses"):
+            raise ValueError(
+                "Distributed cache client cannot decode response, set decode_responses to False"
+            )
+
+    def get(self, *args, **kwargs) -> DistributedCacheReturnType:
+        keyword_args, kwargs, cache_key = self.make_key(args, kwargs)
+        with self.cache_context(cache_key):
+            result = self.client.get(cache_key)
+            if result is None:
+                with self.miss_context(cache_key):
+                    value = self.cached_function(*args, **keyword_args, **kwargs)
+                    self.set(cache_key, value)
+                    return value
+            else:
+                return self.deserialize(result)  # type: ignore
+
+    def set(self, key: str, value: DistributedCacheReturnType) -> None:
+        self.client.sadd(self.get_backend_namespace(), self.namespace)
+        value = self.serialize(value)
+        while self.client.scard(self.namespace) >= self.limit:
+            del_key = self.client.spop(self.namespace)
+            self.client.delete(del_key)
+
+        with self.client.pipeline() as pipe:
+            if self.expire == -1:
+                pipe.set(key, value)
+            else:
+                pipe.setex(key, self.expire, value)
+
+            pipe.sadd(self.namespace, key)
+            pipe.execute()
+
+    def cache_clear(
+        self, args: Optional[tuple] = None, kwargs: Optional[dict] = None
+    ) -> int:
+        if args or kwargs:
+            pattern = self.make_key_pattern(args=args, kwargs=kwargs)
+            delete_keys = list(
+                filter(
+                    pattern.match,
+                    map(bytes.decode, self.client.smembers(self.namespace)),
+                )
+            )
+            if delete_keys:
+                self.client.delete(*delete_keys)
+        else:
+            with self.client.pipeline() as pipe:
+                delete_keys = self.client.smembers(self.namespace)
+                if delete_keys:
+                    pipe.delete(*delete_keys)
+                pipe.delete(self.namespace)
+                pipe.srem(self.get_backend_namespace(), self.namespace)
+                pipe.execute()
+        return len(delete_keys)
+
+    @classmethod
+    def get_all_namespace(cls) -> Set[str]:
+        client = DefaultConfig.get_current_config().cache_redis_client
+        return client.smembers(cls.get_backend_namespace())
+
+    @classmethod
+    def flush_cache(cls) -> int:
+        client = DefaultConfig.get_current_config().cache_redis_client
+        count = 0
+        with client.pipeline() as pipe:
+            for namespace in cls.get_all_namespace():
+                delete_keys = client.smembers(namespace)
+                if delete_keys:
+                    pipe.delete(*delete_keys)
+                    count += len(delete_keys)
+                pipe.delete(namespace)
+            else:
+                pipe.execute()
+        return count
+
+    def serialize(
+        self, value: DistributedCacheReturnType
+    ) -> DistributedCacheReturnType:
+        return value
+
+    def deserialize(
+        self, result: DistributedCacheReturnType
+    ) -> DistributedCacheReturnType:
+        return result