+++ cache_alchemy/backends/pickle.py
@@ -1,0 +1,11 @@
+import pickle
+
+from .base import DistributedCache, BaseCache, ReturnType
+
+
+class DistributedPickleCache(DistributedCache, BaseCache[ReturnType]):
+    def serialize(self, value: ReturnType) -> bytes:
+        return pickle.dumps(value)
+
+    def deserialize(self, result: bytes) -> ReturnType:
+        return pickle.loads(result)