+++ tests/__init__.py
@@ -11,9 +11,7 @@ class TestCacheConfig(DefaultConfig):
 
 def get_config():
     config = TestCacheConfig()
-    config.cache_redis_client = FakeStrictRedis.from_url(
-        config.CACHE_ALCHEMY_REDIS_URL, decode_responses=True
-    )
+    config.cache_redis_client = FakeStrictRedis.from_url(config.CACHE_ALCHEMY_REDIS_URL)
     return config