+++ tests/resource_b.py
@@ -1,6 +1,6 @@
 from unittest.mock import MagicMock
 
-from cache_alchemy import redis_cache, memory_cache
+from cache_alchemy import json_cache, memory_cache
 from tests import get_config
 
 redis_call_mock = MagicMock()
@@ -8,8 +8,8 @@ redis_call_mock = MagicMock()
 config = get_config()
 
 
-@redis_cache
-def redis_test():
+@json_cache
+def json_test():
     redis_call_mock()
     return "test"