+++ tests/test_backends/test_json_cache.py
@@ -5,17 +5,17 @@ from unittest.mock import Mock
 
 from configalchemy.utils import import_reference
 
-from cache_alchemy import redis_cache, method_redis_cache, property_redis_cache
-from cache_alchemy.backends.redis import DistributedCache
+from cache_alchemy import json_cache, method_json_cache, property_json_cache
+from cache_alchemy.backends.json import DistributedJsonCache
 from cache_alchemy.utils import UnsupportedError
 from tests import CacheTestCase
 
 
-class RedisCacheTestCase(CacheTestCase):
+class JsonCacheTestCase(CacheTestCase):
     def test_cache_function(self):
         call_mock = Mock()
 
-        @redis_cache
+        @json_cache
         def add(a: int, b: int = 2) -> int:
             call_mock()
             return a + b
@@ -40,17 +40,17 @@ class RedisCacheTestCase(CacheTestCase):
     def test_cache_namespace_hash(self):
         from tests import resource_a, resource_b
 
-        resource_a.redis_test()
-        resource_b.redis_test()
-        resource_a.redis_test()
-        resource_b.redis_test()
+        resource_a.json_test()
+        resource_b.json_test()
+        resource_a.json_test()
+        resource_b.json_test()
         self.assertEqual(1, resource_a.redis_call_mock.call_count)
         self.assertEqual(1, resource_b.redis_call_mock.call_count)
 
     def test_cache_clear(self):
         call_mock = Mock()
 
-        @redis_cache
+        @json_cache
         def add(a: int, b: int = 2) -> int:
             call_mock()
             return a + b
@@ -66,38 +66,38 @@ class RedisCacheTestCase(CacheTestCase):
 
     def test_cache_set(self):
         self.config.cache_redis_client.flushdb()
-        redis_cache_backend: Type[DistributedCache] = import_reference(
-            self.config.CACHE_ALCHEMY_REDIS_BACKEND
+        json_cache_backend: Type[DistributedJsonCache] = import_reference(
+            self.config.CACHE_ALCHEMY_JSON_BACKEND
         )
 
-        self.assertEqual(set(), redis_cache_backend.get_all_namespace())
+        self.assertEqual(set(), json_cache_backend.get_all_namespace())
 
-        @redis_cache
+        @json_cache
         def add(a: int, b: int = 2) -> int:
             return a + b
 
-        @redis_cache
+        @json_cache
         def mul(a: int, b: int = 2) -> int:
             return a * b
 
         self.assertIn(
-            redis_cache_backend.__name__, add.cache.make_key(args=(1,), kwargs={})[2]
+            json_cache_backend.__name__, add.cache.make_key(args=(1,), kwargs={})[2]
         )
-        self.assertEqual(0, len(redis_cache_backend.get_all_namespace()))
-        self.assertEqual(0, redis_cache_backend.flush_cache())
+        self.assertEqual(0, len(json_cache_backend.get_all_namespace()))
+        self.assertEqual(0, json_cache_backend.flush_cache())
         add(1)
-        self.assertEqual(1, len(redis_cache_backend.get_all_namespace()))
-        self.assertEqual(1, redis_cache_backend.flush_cache())
+        self.assertEqual(1, len(json_cache_backend.get_all_namespace()))
+        self.assertEqual(1, json_cache_backend.flush_cache())
         add(1)
         self.assertEqual(2, add.cache.misses)
         mul(1)
-        self.assertEqual(2, len(redis_cache_backend.get_all_namespace()))
-        self.assertEqual(2, redis_cache_backend.flush_cache())
+        self.assertEqual(2, len(json_cache_backend.get_all_namespace()))
+        self.assertEqual(2, json_cache_backend.flush_cache())
 
     def test_cache_clear_with_pattern(self):
         call_mock = Mock()
 
-        @redis_cache(strict=True)
+        @json_cache(strict=True)
         def add(a: int, b: int = 2) -> int:
             call_mock()
             return a + b
@@ -116,7 +116,7 @@ class RedisCacheTestCase(CacheTestCase):
     def test_strict_cache_function(self):
         call_mock = Mock()
 
-        @redis_cache(strict=True)
+        @json_cache(strict=True)
         def add(a: int, b: int = 2) -> int:
             call_mock()
             return a + b
@@ -141,7 +141,7 @@ class RedisCacheTestCase(CacheTestCase):
         call_mock = Mock()
 
         class Tmp:
-            @method_redis_cache
+            @method_json_cache
             def add(self, a: int, b: int = 2) -> int:
                 call_mock()
                 return a + b
@@ -158,7 +158,7 @@ class RedisCacheTestCase(CacheTestCase):
 
         class Tmp:
             @classmethod
-            @method_redis_cache
+            @method_json_cache
             def add(cls, a: int, b: int = 2) -> int:
                 call_mock()
                 return a + b
@@ -175,7 +175,7 @@ class RedisCacheTestCase(CacheTestCase):
 
         class Tmp:
             @staticmethod
-            @redis_cache
+            @json_cache
             def add(a: int, b: int = 2) -> int:
                 call_mock()
                 return a + b
@@ -193,7 +193,7 @@ class RedisCacheTestCase(CacheTestCase):
 
         class Tmp:
             @property
-            @property_redis_cache
+            @property_json_cache
             def name(self) -> str:
                 call_mock()
                 return name
@@ -203,10 +203,10 @@ class RedisCacheTestCase(CacheTestCase):
         self.assertEqual(Tmp().name, name)
         self.assertEqual(call_mock.call_count, 1)
 
-    def test_redis_cache_limit(self):
+    def test_json_cache_limit(self):
         call_mock = Mock()
 
-        @redis_cache(limit=1)
+        @json_cache(limit=1)
         def add(a: int, b: int = 2) -> int:
             call_mock()
             return a + b
@@ -218,10 +218,10 @@ class RedisCacheTestCase(CacheTestCase):
         self.assertEqual(add(1), 3)
         self.assertEqual(call_mock.call_count, 3)
 
-    def test_redis_cache_expire(self):
+    def test_json_cache_expire(self):
         call_mock = Mock()
 
-        @redis_cache(expire=1)
+        @json_cache(expire=1)
         def add(a: int, b: int = 2) -> int:
             call_mock()
             return a + b
@@ -234,7 +234,7 @@ class RedisCacheTestCase(CacheTestCase):
 
         unexpired_add_call_mock = Mock()
 
-        @redis_cache(expire=-1)
+        @json_cache(expire=-1)
         def unexpired_add(a: int, b: int = 2) -> int:
             unexpired_add_call_mock()
             return a + b