+++ tests/test_dependency.py
@@ -2,7 +2,7 @@ import unittest
 from unittest.mock import Mock
 
 from tests import CacheTestCase
-from cache_alchemy import redis_cache
+from cache_alchemy import json_cache
 from cache_alchemy.dependency import FunctionCacheDependency
 
 
@@ -10,13 +10,13 @@ class DependencyTestCase(CacheTestCase):
     def test_function_dependency(self):
         call_mock = Mock()
 
-        @redis_cache
+        @json_cache
         def add(a, b):
             return a + b
 
         dependency = FunctionCacheDependency(add)
 
-        @redis_cache(dependency=[dependency])
+        @json_cache(dependency=[dependency])
         def add_and_double(a, b):
             call_mock()
             return add(a, b) * 2